modified models.py and made test folder a module
