+++ .travis.yml
@@ -6,6 +6,6 @@ install:
   - pip install -r requirements.txt
 # command to run tests
 script: 
-  coverage run --source Flask_api/ test.py
+  coverage run --source Flask_api/ -m unittest discover
 after_success:
   coveralls
